pyproject.toml
[tool.poetry]
name = "jitter"
version = "0.1.0"
description = "A tool that generates a git branch given a jira ticket number"
authors = ["Jerridan Quiring <jerridan@gmail.com>"]
license = "MIT"
packages = [{include = "main"}]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
Atlassian-Python-Api = "^3.41.3"
GitPython = "^3.1.40"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
jitter = "main.jitter:main"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.2.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
